Fix Errori, Aggiornamento GitIgnore

+++ gdpElaborationApplication/src/main/java/it/univpm/gdpElaborationApplication/dataclass/GDP.java
@@ -2,7 +2,7 @@ package it.univpm.gdpElaborationApplication.dataclass;
 
 public class GDP {
 	private int date;
-	private float value;
+	private double value;
 	
 	public int getDate() {
 		return date;
@@ -10,7 +10,7 @@ public class GDP {
 	public void setDate(int date) {
 		this.date = date;
 	}
-	public float getValue() {
+	public double getValue() {
 		return value;
 	}
 	public void setValue(float value) {
